experimenting with mix level helpers,
for now using a policy that for functions starting with ll_/_ll_ uses the same specialisation rules
as the low-level policy, we probably need user supplied specialisations for the high/mix level helpers themself.

of course if we start introducing new  classes etc, or mixing signatures we will need to be very careful,
and re-invoke as much as necessary of the normalisation logic.

now HintLowLevelOpList has a first sketch of a method .genmixlevelhelpercall(self, function, args_s, args_v)

it is tentatively used to invoke rtimeshift.retrieve_jitstate_for_merge for merge logic blocks.

The dummy code that is there now in rtimeshift.retrieve_jitstate_for_merge is just there for playing with
the concept.



git-svn-id: http://codespeak.net/svn/pypy/dist@23302 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada