rctypes "progress": * union for SomeCTypesObject * prebuilt arrays
Blocker: the skipped test in test_rfunc.py.  The problem is that
(c_char*10) is implemented as an array of structs containing each
a 'value' field of type char.  This makes conversion to a regular
array of chars quite difficult.  I am not sure that there is a
better solution than rewriting rctypes pointers to use addresses
systematically.


git-svn-id: http://codespeak.net/svn/pypy/dist@25695 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada