Refactored exception handling: the first try was to translate Python exception to CLI exceptions, but this would have caused some trobules because the two inheritance hierarchies don't overlap.
Now Python exceptions has its own CLI hierarchy, generated in exactly
the same way that is for other classes. This means that we can no
longer rely on CLI '.ovf' opcodes that raises OverflowException on
error, because Python's OverflowError doesn't map to CLI's
OverfloweException. For now the tricky and inefficient solution is to
explicitly catch OverflowException and throw a fresh OverflowError
instead.



git-svn-id: http://codespeak.net/svn/pypy/dist@25725 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada