fix propagation of exception out r_dict operations.
Notice that this required having the flow space assume that getitem/setitem/delitem can raise any exceptions.
Changes with some tests at various levels to cope with this.

Notice that the pruning of excepts links when a getitem appears in an unrelated try/except is less strong as a 
consequence. 

Now {}[[]] fails with a TypeError as expected in a translated PyPy.



git-svn-id: http://codespeak.net/svn/pypy/dist@25965 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada