(pedronis, arigo)
Yikes!  The problem was about the stackless transform.  It contained a hard-coded
list of operations that can possibly unwind the stack.  Moreover, even if it knew
about the new gc_x_*() operations, the exception transform would kill the
UnwindException branch without a canraise=... in LLOp().  So we added a flag
'canraisegc=True' that prevents both things from going wrong.

test_tree_cloning passes now!  Cloning seems to be working well enough that we
can try to expose it to RPython and see.



git-svn-id: http://codespeak.net/svn/pypy/dist@27519 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada