Trying to sort out the dependencies in order between the stackless and the gc transforms.  The isgchelper flag was getting in the way because gc helpers must almost all be stackless transformed too (e.g. to allow UnwindException to pass through a call to a destructor).
Some obscure hacks now allow cooperation between the mixlevelannotator
and the genc database about delayed functions.  It is now ok to get()
delayed function pointers.  It forces the unique pypy_g_* name of the
function to be computed, but the FuncNode is not built.  Instead, the
complete() method periodically checks if these delayed pointers have
been resolved (e.g. by a finish()).

The finish() of the gc transformer is now in two parts: finish_helpers()
finishes the mixlevelannotator; then finish_tables() computes the
final GC tables.

Documented in complete() the precise order in which these finish() are
called and why.


git-svn-id: http://codespeak.net/svn/pypy/dist@27622 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada