some of the flags in the os module such as os.O_CREAT & co. are simply different depending on the platform: this means that the same rpython program can be different depending on the platform it is *compiled* on. Moreover, these values are passed to external functions implemented in pypylib.dll, which has no chance to know the original meaning of the flags.
We really don't want this because .NET executables should be portable
between windows and Mono: to solve, we patch some of the attributes of
the os module to have the same value as in Windows just before the
annotation, then we restore the original values when the compilation
has been completed.

This could be dangerous if "real" python code (instead of rpython
code) tries to access such attributes when they are patched, but as
far as now tests are still passing, so I hope not have broken anything.

This is only a short-term hack, the long term solution will be to
refactor the I/O model of pypy.



git-svn-id: http://codespeak.net/svn/pypy/dist@30960 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada