(pedronis, arigo)
Pass an extra test about catching exceptions in JITted code.

More and more based on the exceptiontransform, this required some
clean-ups and a strange flag to the latter to produce graphs that give
more hints to the JIT.

Added a sanity-check to llinterp: when returning from
exceptiontransformed graphs with an exception set, the return value
should really be the special error value.



git-svn-id: http://codespeak.net/svn/pypy/dist@32035 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada