Add support for optional __slots__ in the annotator.  If an attribute appears on a class, but is not in the __slots__ of that class, then there is probably a problem: the annotator moved the attribute up to a common base class when you were not expecting it.
(Just to make sure this is clear, our usage of __slots__ is not about
what attributes *should* appear, and certainly not about giving type
hints.  The annotator doesn't mind if a program doesn't seem to be using
all the attributes specified in the __slots__.  Such attributes won't
appear in compiled programs anyway.  And you couldn't tell the
difference anyway :-)

(And don't go adding __slots__ everywhere please!  It's for abstract
base classes kind of usage.)



git-svn-id: http://codespeak.net/svn/pypy/dist@32088 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada