killing builtin frames as such. first pass at killing the convoluted setup in gateway.py.
Still using instances called activations to share code for similar patterns of unwrap specs: whether is worth
to be checked later by trying to go for just attaching directly a "run" function to the builtin code object.

PyPy is slightly smaller.

The emit recipe can/needs to be simplified further.



git-svn-id: http://codespeak.net/svn/pypy/dist@32544 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada