Another (better?) start at the socket module.  This one is meant to be usable from pure RPython programs (methods with a 'space' arguments are only to ease the implementation of the future PyPy mixed module).  In particular it should work well on top of CPython too.
The address-manipulation parts of the interface are not like CPython's;
instead, to be RPythonic, we have various subclasses of a base Address
class, with various constructors and inspection methods.  At least this
allows some roughly clean separation of the manipulation of the various
address families.

Implementation-wise, this tries to stick as close to CPython's
socketmodule.c as possible.



git-svn-id: http://codespeak.net/svn/pypy/dist@32598 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada