(arre, pedronis)
support specifying global merge points using a hint(None, global_merge_point=True). The block of the hint if it's a merge block,
or the previous block, or exactly one of the following blocks will be considered otherwise depending whether the hint is the first
or the last operation in the block. In other cases the situation will be considered ambiguous.



git-svn-id: http://codespeak.net/svn/pypy/dist@33912 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada