issue266 testing
A rewrite of the compiler's scoping logic.  Trying to be more explicit
about the rules this time...  The diff certainly contains a lot of
*removed* lines, in favor of a more centralized logic.  This version
also solves a few bugs that we had (see new tests), and solves a couple
of problems in CPython with class vs. scope interaction.

The main new incompatibilities with CPython are:

1. some code involving classes and import * / exec now raise a
  SyntaxError instead of producing broken bytecode, which I think is 
  not a bad thing (there are pending SourceForge bug reports for CPython)

2. in class scopes, variables that just go "through" (because they are
  free in a method and come from an enclosing function) now have their
  name mangled, if you try to inspect them with locals().  This avoids
  confusion between a class-level name and the variable going "through".
  It fixes another CPython bug (also reported on SF I think) which
  allows strange code to modify bindings in parent scopes.

This fully passes CPython's test_scope, and all our previous tests
with the exception of the cases now rejected by 1. above.


git-svn-id: http://codespeak.net/svn/pypy/dist@34441 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada