Added support for handling native exceptions when using .NET-types.
It's a bit involved because .NET exceptions don't share a common root
with RPython exceptions, so we need to cheat a bit by creating a fake
class that will be recognized by the backend and rendered as
appropriate.

This means, for example, that instead of writing something like this:

try: 
    f()
except CLR.System.Exception:
    ...

we need to create the fake object first.

SystemException = NativeException(CLR.System.Exception)

try:
    f()
except SystemException:
    ...

Please note that the call to NativeException is *not* RPython, so you
need to call it at initialization-time.

Thanks to pedronis for the original idea.



git-svn-id: http://codespeak.net/svn/pypy/dist@34602 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada