(pedronis, arre on the earlier sweat part, arigo)
Don't use the RPython's normalization in the hint-annotator.  It took
quite some sweat to discover that it was not a good idea.  Instead, for
all the variables whose color is not fixed at the end of
hint-annotation, we proceed by assuming that they are all green and
forcing them to red as needed, in a fixpoint process again.  The "needs"
is based on the dependencies captured by 'myorigin' and by the call
families, which produces the normalization effect.


git-svn-id: http://codespeak.net/svn/pypy/branch/jit-real-world@35278 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada