(pedronis, arigo)
Whack whack whack.  'do_imports_immediately' makes no sense as a
translation config, because sub-annotations with different policies
(e.g. the annlowlevel ones) must use the default True, or they crash.
So move it as an annotator policy flag.  But then geninterp has no
annotator policy, so hack a bit more for that case.

Fun fun fun.

Fix some more stuff in test_wrapping and raymond.  Fix c/wrapper to cope
with the extra genop checks.  Fix something that broke because of that.
It also seems that our default gcpolicy is not refcounting any more, not
sure why.  Show the space operation in the "annotation degenerated to
SomeObject()" message.



git-svn-id: http://codespeak.net/svn/pypy/dist@35394 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada