add the possibility to export classes.
A class is exported is its __init__ is marked with @export. Individual
methods can also be @exported, useful for methods that can't be
reached by one of the entrypoints.

To force the annotator to analyze a given method, we generates a dummy
function which call that method. Better approached are welcome.



git-svn-id: http://codespeak.net/svn/pypy/dist@44557 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada