issue325 testing
It's too easy to use id() in RPython and forget that it's actually a
pretty advanced operation on top of moving GCs.  This replaces it with
two functions in the objectmodel module, one which implements the full
semantics of id() and another which just returns the current address as
an integer.


git-svn-id: http://codespeak.net/svn/pypy/dist@49232 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada