Merge of the lazy-write-barrier branch by fijal.
With the generation gc, prebuilt gc objects are now only recorded as
potential static roots when a gc pointer is written into them at
run-time.  This reduces the number of static roots a lot, although
it only gives marginal speed improvements (there seem to be not that
many full collects during a regular pypy-c execution).



git-svn-id: http://codespeak.net/svn/pypy/dist@49850 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada