Passing structure as values to functions.
_ffiletter is probably not anymore the right name for this.

This leaks ffitypes at the moment, given that for some ABIs we really need
the completely filled ffitype we may want to approach this by having gettypecode
return something that contains a full ffitype for the structure(/array) type,
instead of the current tuple. The type would own its ffitype then.



git-svn-id: http://codespeak.net/svn/pypy/dist@51791 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada