port of fixes from the smalltalk branch:
------------------------------------------------------------------------
r55149 | cfbolz | 2008-05-23 16:22:38 +0200 (Fri, 23 May 2008) | 2 lines
Changed paths:
   M /pypy/branch/smalltalk-shadow-changes/pypy/translator/backendopt/malloc.py
   M /pypy/branch/smalltalk-shadow-changes/pypy/translator/backendopt/test/test_malloc.py

a tentative fix for the malloc removal problems

------------------------------------------------------------------------
r55120 | cfbolz | 2008-05-22 09:20:42 +0200 (Thu, 22 May 2008) | 5 lines
Changed paths:
   M /pypy/branch/smalltalk-shadow-changes/pypy/translator/backendopt/test/test_malloc.py

failing test about malloc removal in the presence of cast_pointers. The
problem is that malloc removal replaces each occurence of the same
structure in the input args of a block by the exploded vars, which means
that there is no global view of what the variables currently are.



git-svn-id: http://codespeak.net/svn/pypy/dist@55334 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada