The cycle breaking algorithm is broken: it doesn't break all cycles.  To show this I added "assert is_acyclic()" at the end of break_cycles().
A simple fix, namely "repeat break_cycles() until is_acyclic()",
requires more than 7 minutes to compute on the pypy call graph again.
So this check-in contains an experimental tweak that reduces the run
time to about 1 minute.  I am quite unsure about the quality of the
result but I am roughly guessing that it is similar to the
7-minutes-running version.


git-svn-id: http://codespeak.net/svn/pypy/dist@55955 fd0d7bf2-dfb6-0310-8d31-b7ecfe96aada